Changed the sportsBook file to the main menu
This is the new main menu system. It shouldn't need any work for a while as it does everything I wanted it to do. 
It will have to import the main function for each additional sport (like I said, a while away!).

The previous football script is now football.py.