The original sportsBook script
This was originally the main file, but that is now the menu. This is currently the main football file.

This fill is messy and I've learned a lot since writing it. It's still the basis of my project, but I'll probably split some of it, and add some more, into additional files.

One big step will be to add an appropriate football menu. The user will have a choice of uploading leagues, running comparisons and running reports. I have plans for reports, but that's for another file description!