Bug fix
Fixed a tiny bug caused by the last edit (change of text in a menu item). The problem was caused by the menu item being recognised by its name. I've changed the text in the checking if statement for now, but this isn't an ideal system. Something for review at a later date.