Update sportsBook.py
Added author information.

Added leagueData initialisation (moved from the football menu file).

Broke the menu system! (Intentionally!) - The previous menu was nice and tidy but meant that associated functions had to be called in the same way. In reality, I need some functions to assign a value to an item and some just to display information or write it to a file. The current system is much more "manual", but gives the freedom I need.