Update footballMenu.py
Menu now handles predictions that are returned from game analysis functions (the only current such function is the manualGameAnalysis function). Returned predictions are added to a list of predictions. For manual game analysis, the option is now provided to run analysis on further teams.

In the reports menu there is now an option to display current predictions.