Improved team display
Following on from Surister's awesome work, I've tweaked the
custom_pretty_print function to work with lists too.
I've modified the code used to display the available teams in the manual analysis menu so that it produces a nice tidy list for displaying with the custom_pretty_print function.

I've also updated a couple of docstrings and corrected a typo.