Create league_and_team_classes.py
I know simplicity was the main aim in creating OOC. I've created complication in linking two classes. However, it means that the League class looks after itself. For every Team object instantiated, a new League is created if necessary. Each League object will have it's teams added as they are scraped.

As a result, the classes aren't as simple as they could be. Hopefully not a problem to the more experienced programmer. Although, it doesn't really matter as long as it is understood what it does.