Upload of experiment file - sample menu class
This file is a sample of how a single menu class can be used to create all menus. This, or something like this will significantly reduce the amount of code used and make tweaking the code much easier. It'll also make the projcet appear more uniformed throughout.

I've also created a placeholder footballMenu file so that the main_menu object runs.