#36 and more
Moved the import and export functions to common functions.

Changed them so that they ask for a file name when saving and loading.

Updated the prediction list to store the league name.

Updated the predicitons display so that leagues are tidily separated and titled.

Noted the export league data doesn't export fixtures (so it can't be used to run automatic predictions - for now).

Added more leagues.