Added multiple input for league selection.
Also added a bit of code to make it easier to add extra options to the league selection code.

Added an "all" option to scrape all data.

Renamed the display_selection function to inform_and_scrape.

Changed the information displayed when downloading data to suit the new method.

Updated a few docstrings.