Write to Excel - inprogress ++
I've edited and renamed the export_to_json function so that it can now, to some degree, export to Excel. It takes parameters to choose either json or excel for saving. It can also take "choose" and present the user with an option.

I've added another variable to the predictions list. prediction_goal_difference holds the absolute difference between the predicted home and away scores.

Corrected a team count error in the available_leagues dictionary.