Available leagues now stored in an external file.
Simplified the available leagues dictionary. Now it only needs the league name and link. League option numbering is done by by code as is calculating the number of teams in each league.

The available_leagues dictionary is now stored in a json file. This means users can edit it without needing to edit any code.

Added 9 more leagues to the avalable leagues file.

Sorted the leagues into alphabetical order of country for ease of locating (and checking for duplicates).

Tweaked the import_json_file function so that it can load a specified file name without prompting for it and can do so silently if requested to do so.

Tidied a couple of docstrings and comments.