Better error handling when site is down
When the betstudy site is down, a response is still provided. This means that typical error catching methods don't work. The last update saw a catch for this type of error, but, as the site was up and running I was unable to test it.

It's down again today so I've tweaked and tided the displayed messaged. I've also prevented the manual game analysis option from asking if you'd like to analyse another game when it failed on the previous attempt due to no league information being available.