Datetime object implementation
Created date time objects in the soccerstats scraper which are returned along with a tidy representation of the date and time of each game. For later comparison.

Also bypassed the initial menu as it's a development placeholder. A comment in the appropriate place of sportsBook.py explains how to reactivate the menu.