Implemented datetime objects to betstudy scraper
The bet study scraper now returns an appropriately formatted date with it's fixtures and a datetime object too. This will be used for specifying the range of fixtures to analyse later.

Completed the bypassing of the sport select menu so that users now quit from the football menu. The previous code is commented out for future reimplementation.