Added game range options to reports + bug fixes
Fixed a bug that allowed an empty list of fixtures to be displayed.

Fixed a bug that allowed an empty dictionary league data to be displayed.

Tidied the reports menu so that it is properly displayed when returning from a sub menu.

Implemented the menus and variables for selecting the game range (Yet to be put into action).

Added a notification to advise when predictions have been processed.