Bug fix
Bug fix: Some fixtures were being identified as "Mixed League" because a function to determine the league a team belonged to didn't cope with the same team name existing in multuple leagues.

The problem is fully resolved for automatic predictions based on fixtures. The problem has been significantly reduced for manually entered games. The function now checks the same league first. It the two teams can't be found in the same league, the first matching leagues are returned. There may be some very rare and unlikely cases where the wrong league may be selected, resulting in inaccurate statistics. The bug remains that the team isn't being added to the list with a new id number when a team of the same name already exists.

Tidied the "Manual Entries" heading in display_predictions.

Fixed a bug that meant some teams could be predicted to score half goals.