Added prediction filtering
Added filtering options to the reports menu. This enables the user to filter types of expected outcome from the generated predictions. Eg, games where the home team are expected to win by 3 or more goals, or games where a total of 5 goals are expected.

Added "Total goals expected"  and "Total goals scored" to the prediction dictionary and as a column to the prediction spreadsheet.