Bug fixes and minor improvements
Fixed a bug that caused the script to hang when there was a problem scraping data.

Updated scraper error messages to advise the problem could be caused by leagues being in play off or group stages.

Reduced the number of threads used when scraping soccerstats as further threads appear to provide no benefit and may cause the site unnecessary traffic.

Slightly changed the method used to prevent division by zero in the scrapers to more efficient method (thanks Surister)

When filtering predictions a message will now advise you when applying a filter will leave no games remaining and will prevent the filter from being applied.

Option 9 in the main menu now shows that it will switch the data source to the next source name.

Fixed a bug in one of the filters which was referencing on old key name in predictions.

When a filter is applied to predictions a check is now done to see if the filter has already been applied. If so the user is advised and no action is taken.

Clarified the win by x goals filter text and fixed a bug in the code that would have prevented some games from showing.