Bet Study results scraper + Problem identified on soccerstats scraper.
Identified a problem with the SoccerStats scraper that meant some leagues weren't being scraped due to an alternative web page format in use. Fixed the result scraping side of that. Still need to fix the fixture scraping side of it. Example is the Argentinian Primera league. This problem does not exist with the BetStudy scraper. Fixing this is my priority!

Completed the BetStudy results scraper. This means that development with results based algorithms can begin once the above problem is resolved.

Fixed a bug in the team to score x or less predictions filter which caused teams to score x or more to be saved instead.

Removed the text advising that entering the Filter Predictions menu clears all predictions as it no longer does.