Results system and more
Implemented results filtering by date with a few new functions to deal with them.

Moved the visual_comparison function to a new visuals.py file. The file will be used for all visual representations.

Renamed the game_range object to future_range so that it can be related to the newly required past_range(for results).

Tidied the reports menu (slightly).

Updated most of the docstrings.

I've ran quite a few tests and all seems well. Please report any issues!