New feature + minor fixes and updates
Added an awesome new option: Benchmark Analysis
This allows the user to select an upcoming fixture and see what common opponents they have in their results this season.
It currently shows the results but more options will follow.

Removed some outstanding debgugging code in the display results in range function.

Changed the football menu so that fixtures in range and results in range are created as soon as a scrape is complete.

The select league screen now advises what leagues have been selected as they have been selected.