New feature: Predictions by benchmark data only + more
Added the ability to run a full set of predictions based on benchmark game data alone.
This means there is a new option when rrunning predictions.

Moved the predictive algorithms to a separate file.

Fixed issue #84 (prevented old, unplayed fixtures from showing in the fixtures list).
Fixed a newly found bug which caused a crash in benchmark analysis when one of the previously played games had been postponed.