New filter option + fixes
Moved some code from filteredPredictions into commonFunctions as callable functions.
Corrected the description of the James Shoemark Special filter.
Added some new filters that filter predictions based on previous results (in the defined results range).
Fixed a bug that affected the results returned when the selected range was a number of games.
Fixed a bug that affected the fixtures returned when the selected range was a number of games.
Improved to displaying of results and fixtures so that they are ordered by league rather than randomly scattered.
Added templates for more form filter options to be added in the future.
Edited the scrapers/py file so that failed scrapes are clearly identified.
Removed Serbia - Prva Liga from the SoccerStats list (json file) as the league is no longer on the site.