Finished the form filter + Isue #84 + function improvement
Changed the input_goals function in commonFunctions.py to input_number. It now just takes a number, validates it and returns it as an int.
Improved the form predictions filter so that it now asks for the range of results to consider.
Improved the formatting of relevant results when they are requested to be displayed in the form predictions filter.
Set the default results range to 365 days.
Finally fixed Issue #84 which was causing games with past dates to show in predictions.